Add V1 contracts

See merge request pantherprotocol/panther-core!2488

+++ contracts/contracts/protocol/v1/mocks/MockTransactionNoteEmitter.sol
@@ -2,14 +2,14 @@
 // SPDX-FileCopyrightText: Copyright 2021-25 Panther Protocol Foundation
 pragma solidity ^0.8.19;
 
-import "../pantherPool/TransactionNoteEmitter.sol";
+import "../core/utils/TransactionNoteEmitter.sol";
 
 contract MockTransactionNoteEmitter is TransactionNoteEmitter {
     event LogPrivateMessage(bytes privateMessage);
 
     function internalSanitizePrivateMessage(
         bytes memory privateMessages,
-        uint8 txType
+        uint16 txType
     ) external {
         _sanitizePrivateMessage(privateMessages, txType);
